<==== Mercurial data ====>
Branch: x_CLOUD-593_rename_api_camelCase

+++ client/libclient/src/ui/widgets/resource_preview_widget.cpp
@@ -15,7 +15,7 @@
 
 namespace {
 
-    const int kDefaultAspectRatio = 4.0 / 3.0;
+    const qreal kDefaultAspectRatio = 4.0 / 3.0;
 
 } // namespace
